Tighten compilation flags in Emscripten builds
* Add "-Werror" - i.e., treat warnings as compilation errors. (To note,
  in Native Client builds this has always been enabled by NaCl's
  toolchain.)
* Enable "NDEBUG" preprocessor definition in Release builds, in order to
  disable debug-only assertions and debug value output. (In NaCl builds
  this has already been enabled.)
* Fix a few warnings detected by Emscripten's clang compiler:
* * logging.cc:35:16: error: unused variable 'kTypeMessageKey'
* * logging.cc:36:16: error: unused variable 'kMessageType'
* * logging.cc:37:16: error: unused variable 'kDataMessageKey'
* * logging.cc:38:16: error: unused variable 'kDataLogLevelMessageKey'
* * logging.cc:39:16: error: unused variable 'kDataTextMessageKey'
* * value_unittest.cc:80:38: error: implicit conversion from 'const int64_t'
    (aka 'const long long') to 'double' changes value.

+++ common/cpp/src/google_smart_card_common/value_unittest.cc
@@ -77,7 +77,7 @@ TEST(ValueTest, Integer64BitMax) {
   EXPECT_FALSE(value.is_dictionary());
   EXPECT_FALSE(value.is_array());
   EXPECT_EQ(value.GetInteger(), integer_value);
-  EXPECT_DOUBLE_EQ(value.GetFloat(), integer_value);
+  EXPECT_DOUBLE_EQ(value.GetFloat(), static_cast<double>(integer_value));
 }
 
 TEST(ValueTest, Integer64BitMin) {
